[IW-24] Add flask server handling

+++ backend/scripts/audio_splitter.py
@@ -1,9 +1,10 @@
-import argparse
 import os
 import subprocess
 import sys
 from audio_extract import extract_audio
 from pytubefix import YouTube
+
+resources_path = "src/main/resources"
 
 
 def splitAudio(song_file, directory):
@@ -17,15 +18,16 @@ def splitAudio(song_file, directory):
 
 def download_audio_from_youtube_url(youtube_url, directory_name):
     try:
-        songs_dir = f"songs/{directory_name}"
+        songs_dir = f"{resources_path}/songs/{directory_name}"
         os.makedirs(songs_dir, exist_ok=True)
         yt = YouTube(youtube_url)
         ys = yt.streams.filter(only_audio=True).first()
-        file_path = ys.download(output_path="./" + songs_dir, filename=directory_name + ".mp4")
 
-        file_name = os.path.relpath(file_path)
         output_wav = songs_dir + "/" + directory_name + ".wav"
         if not os.path.exists(output_wav):
+            file_path = ys.download(output_path=f"./{songs_dir}",
+                                    filename=f"{resources_path}/{directory_name}.mp4")
+            file_name = os.path.relpath(file_path)
             extract_audio(input_path=file_name, output_path=output_wav,
                           output_format="wav")
             os.remove(file_name)
@@ -41,7 +43,7 @@ def extract_video_id(youtube_url):
     return video_id
 
 
-def wavToMp3(fileName):
+def wavToMp3(fileName, removeWav=True):
     wavFile = fileName + ".wav"
     mp3File = fileName + ".mp3"
 
@@ -49,30 +51,31 @@ def wavToMp3(fileName):
         extract_audio(input_path=wavFile, output_path=mp3File,
                       output_format="mp3")
 
-    if os.path.exists(wavFile):
+    if removeWav and os.path.exists(wavFile):
         os.remove(wavFile)
 
+    if os.path.exists(mp3File):
+        return mp3File
+    else:
+        raise
 
-def main():
-    parser = argparse.ArgumentParser(description="Download and process audio from YouTube.")
-    parser.add_argument("youtube_url", type=str, help="The URL of the YouTube video to download.")
 
-    args = parser.parse_args()
+def split(youtube_url):
+    directory_name = extract_video_id(youtube_url)
+    songs_dir = f"{resources_path}/songs/{directory_name}"
 
-    directory_name = extract_video_id(args.youtube_url)
-    songs_dir = f"songs/{directory_name}"
-
-    file_name = download_audio_from_youtube_url(args.youtube_url, directory_name)
+    file_name = download_audio_from_youtube_url(youtube_url, directory_name)
 
     if file_name:
-        splitAudio(file_name, songs_dir)
-
         vocals = f"{songs_dir}/{directory_name}_Vocals"
         instruments = f"{songs_dir}/{directory_name}_Instruments"
 
-        wavToMp3(vocals)
-        wavToMp3(instruments)
+        if not os.path.exists(vocals + ".mp3") or not os.path.exists(vocals + ".mp3"):
+            splitAudio(file_name, songs_dir)
 
-
-if __name__ == "__main__":
-    main()
+            wavToMp3(vocals)
+            wavToMp3(instruments)
+    return {
+        "vocalsPath": f"/api/audio/split/{directory_name}/{directory_name}_Vocals.mp3",
+        "instrumentsPath": f"/api/audio/split/{directory_name}/{directory_name}_Instruments.mp3"
+    }
